Wed, Jul 27, 2022, 12:26 PM +00:00

+++ weeks/week2/hi.c
@@ -6,5 +6,5 @@ int main(void)
     char c2 = 'I';
     char c3 = '!';
 
-    printf("%i %i %i\n", c1, c2, c3);
+    printf("%i %i %i\n", (int) c1, (int) c2, (int) c3);
 }
